+++ pharmacy/models.py
@@ -1,5 +1,6 @@
 from django.db import models
-
+from user.models import Users
+from pill.models import Vitamins
 
 class Pharmacy(models.Model):
     """
@@ -27,3 +28,28 @@ class Pharmacist(models.Model):
     description = models.CharField(max_length=50, null=True, blank=True)
     age = models.IntegerField(null=True, blank=True)
 
+class Pharmacy_Review(models.Model):
+    """
+    약국 리뷰 정보에 대한 모델
+    """
+    id = models.AutoField(primary_key=True)
+    user_id = models.ForeignKey(Users, on_delete=models.CASCADE)
+    pharmacy_id = models.ForeignKey(Pharmacy, on_delete=models.CASCADE)
+    rating = models.SmallIntegerField()
+    content = models.CharField(max_length=50, null=True)
+
+class PharmacyLikes(models.Model):
+    id = models.AutoField(primary_key=True)
+    user_id = models.ForeignKey(Users, on_delete=models.CASCADE)
+    pharmacy_id = models.ForeignKey(Pharmacy, on_delete=models.CASCADE)
+
+class Orders(models.Model):
+    id = models.AutoField(primary_key=True)
+    user_id = models.ForeignKey(Users, on_delete=models.CASCADE)
+    pharmacy_id = models.ForeignKey(Pharmacy, on_delete=models.CASCADE)
+    created_at = models.DateTimeField(auto_now_add=True)
+
+class PharmacyVitamins(models.Model):
+    id = models.AutoField(primary_key=True)
+    vitamin_id = models.ForeignKey(Vitamins, on_delete=models.CASCADE)
+    pharmacy_id = models.ForeignKey(Pharmacy, on_delete=models.CASCADE)
