+++ user/models.py
@@ -1,5 +1,4 @@
 from django.db import models
-from pharmacy.models import Pharmacy
 from pill.models import Vitamins
 
 
@@ -12,11 +11,6 @@ class Users(models.Model):
     location = models.CharField(max_length=50, null=True)
 
 
-class PharmacyLikes(models.Model):
-    id = models.AutoField(primary_key=True)
-    user_id = models.ForeignKey(Users, on_delete=models.CASCADE)
-    pharmacy_id = models.ForeignKey(Pharmacy, on_delete=models.CASCADE)
-
 
 class VitaminLikes(models.Model):
     id = models.AutoField(primary_key=True)
@@ -28,11 +22,4 @@ class BuyList(models.Model):
     id = models.AutoField(primary_key=True)
     user_id = models.ForeignKey(Users, on_delete=models.CASCADE)
     pharmacy_id = models.ForeignKey
-
-
-class Orders(models.Model):
-    id = models.AutoField(primary_key=True)
-    user_id = models.ForeignKey(Users, on_delete=models.CASCADE)
-    pharmacy_id = models.ForeignKey(Pharmacy, on_delete=models.CASCADE)
-    created_at = models.DateTimeField(auto_now_add=True)
